Updated README.md
This is the initial draft of the README document. While writing the README, we are going to need to separate this application into two smaller applications. The first application is going to be for the owner of the gym. The second application is going to be for the members of the gym.